+++ jest.config.js
@@ -1,8 +1,9 @@
 module.exports = {
-  "roots": [
-    "<rootDir>/test"
-  ],
-  "transform": {
-    "^.+\\.ts$": "ts-jest"
+  collectCoverage: true,
+  verbose: true,
+  collectCoverageFrom: ['src/**/*.ts'],
+  transform: {
+    "^.+\\.tsx?$": "ts-jest",
   },
+  testRegex: "(./tests/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
 }