+++ src/domain/builder/builder.ts
@@ -1,5 +1,7 @@
-import { IBuilder, IFile, IStore } from "../../interfaces";
-import { cssBuilder, htmlBuilder } from "./builders";
+import { IBuilder, IFile, ICache } from "../../interfaces";
+import { cssBuilder, htmlBuilder, markdownBuilder } from "./builders";
+import { join } from "path";
+import { DirScanner } from "../../lib/dir-scanner";
 
 export interface IBuilders {
   [key: string]: IBuilder;
@@ -7,21 +9,39 @@ export interface IBuilders {
 
 export const defaultBuilders = {
   html: htmlBuilder,
-  css: cssBuilder
+  css: cssBuilder,
+  md: markdownBuilder
 };
 
 export class Builder {
-  private store: IStore;
+  private cache: ICache;
   private builders: IBuilders;
 
-  constructor(store: IStore, builders: IBuilders) {
-    this.store = store;
+  constructor(cache: ICache, builders: IBuilders) {
+    this.cache = cache;
     this.builders = builders;
   }
 
   public async buildFile(file: IFile) {
     const builder = this.builders[file.extension];
-    if (builder) return builder(this.store, file);
+    if (builder) return builder(this.cache, file);
     else return file;
+  }
+
+  public async build() {
+    const config = this.cache.get("config");
+    const publicFolder = join(config.projectFolder, config.publicFolder);
+    const posts = this.cache.get("posts").map((p: IFile) => ({
+      ...p,
+      path: config.postsFolder
+    }));
+
+    const publicFiles = await Promise.all(
+      [...(await DirScanner.scanAndGetFiles(publicFolder)), ...posts].map(pf =>
+        this.buildFile(pf)
+      )
+    );
+
+    return publicFiles;
   }
 }