+++ src/domain/loader/index.ts
@@ -1,2 +1,1 @@
-export * from "./filler";
 export * from "./loader";