+++ test/unit/builder.html.test.ts
@@ -1,11 +1,11 @@
-import { Store } from "../../src/lib/store";
+import { MemoryCache } from "../../src/lib/cache";
 import { IFile } from "../../src/interfaces";
 import { htmlBuilder } from "../../src/domain/builder/builders";
 import { DirScanner } from "../../src/lib/dir-scanner";
 
 describe("Builder - HTML", () => {
   test("Builds file without template", async () => {
-    const store = new Store();
+    const cache = new MemoryCache();
     const fakeFile: IFile = {
       name: "article.html",
       extension: "html",
@@ -19,13 +19,13 @@ describe("Builder - HTML", () => {
       raw: " <p>test</p>"
     };
 
-    const output = await htmlBuilder(store, fakeFile);
+    const output = await htmlBuilder(cache, fakeFile);
     expect(output).toMatchObject(result);
   });
 
   test("Template not found", async () => {
-    const store = new Store();
-    store.set("templates", {});
+    const cache = new MemoryCache();
+    cache.set("templates", {});
     const fakeFile: IFile = {
       name: "article.html",
       extension: "html",
@@ -34,14 +34,14 @@ describe("Builder - HTML", () => {
       raw: "<!--\n @template main \n--> <p>test</p>"
     };
 
-    return expect(htmlBuilder(store, fakeFile)).rejects.toEqual(
+    return expect(htmlBuilder(cache, fakeFile)).rejects.toEqual(
       new Error("Template main not found.")
     );
   });
 
   test("Replace content correctly", async () => {
-    const store = new Store();
-    store.set("templates", {
+    const cache = new MemoryCache();
+    cache.set("templates", {
       main: "<div>{{content}}</div>"
     });
 
@@ -58,17 +58,17 @@ describe("Builder - HTML", () => {
       raw: "<div> <p>test</p></div>"
     };
 
-    const output = await htmlBuilder(store, fakeFile);
+    const output = await htmlBuilder(cache, fakeFile);
     expect(output).toMatchObject(result);
   });
 
   test("Replace snippets inside template correctly", async () => {
-    const store = new Store();
-    store.set("config", { mode: "prod" });
-    store.set("templates", {
+    const cache = new MemoryCache();
+    cache.set("config", { mode: "prod" });
+    cache.set("templates", {
       main: "<div>{{snippet:myvar1}}{{content}}</div>"
     });
-    store.set("snippets", {
+    cache.set("snippets", {
       myvar1: { configMode: "all", value: "ABCD" },
       myvar2: { configMode: "all", value: "XYZA" }
     });
@@ -86,17 +86,17 @@ describe("Builder - HTML", () => {
       raw: "<div>ABCD <p>XYZA</p></div>"
     };
 
-    const output = await htmlBuilder(store, fakeFile);
+    const output = await htmlBuilder(cache, fakeFile);
     expect(output).toMatchObject(result);
   });
 
   test("Replace snippets configMode", async () => {
-    const store = new Store();
-    store.set("config", { mode: "prod" });
-    store.set("templates", {
+    const cache = new MemoryCache();
+    cache.set("config", { mode: "prod" });
+    cache.set("templates", {
       main: "<div>{{snippet:myvar1}}{{content}}{{snippet:myvar3}}</div>"
     });
-    store.set("snippets", {
+    cache.set("snippets", {
       myvar1: { configMode: "prod", value: "ABCD" },
       myvar2: { configMode: "dev", value: "XYZA" },
       myvar3: { configMode: "all", value: "always" }
@@ -115,17 +115,17 @@ describe("Builder - HTML", () => {
       raw: "<div>ABCD <p></p>always</div>"
     };
 
-    const output = await htmlBuilder(store, fakeFile);
+    const output = await htmlBuilder(cache, fakeFile);
     expect(output).toMatchObject(result);
   });
 
   test("No recent post template file", async () => {
-    const store = new Store();
-    store.set("config", { recentPosts: 2 });
-    store.set("templates", {
+    const cache = new MemoryCache();
+    cache.set("config", { recentPosts: 2 });
+    cache.set("templates", {
       main: "<div>{{content}}</div>"
     });
-    store.set("posts", DirScanner.scanAndGetFiles("./test/data/project/posts"));
+    cache.set("posts", DirScanner.scanAndGetFiles("./test/data/project/posts"));
 
     const fakeFile: IFile = {
       name: "article.html",
@@ -135,20 +135,20 @@ describe("Builder - HTML", () => {
       raw: "<!--\n @template main \n--> <div>{{blog:recent-posts}}</div>"
     };
 
-    return expect(htmlBuilder(store, fakeFile)).rejects.toThrow(
+    return expect(htmlBuilder(cache, fakeFile)).rejects.toThrow(
       new Error("There is no template for recent posts.")
     );
   });
 
   test("Recent posts", async () => {
-    const store = new Store();
-    store.set("config", { recentPosts: 2 });
-    store.set("templates", {
+    const cache = new MemoryCache();
+    cache.set("config", { recentPosts: 2 });
+    cache.set("templates", {
       main: "<div>{{content}}</div>",
       recentPost:
         "<div>{{title}}</div><p>{{author}}</p><p>{{description}}</p><p>{{date}}</p>"
     });
-    store.set(
+    cache.set(
       "posts",
       await DirScanner.scanAndGetFiles("./test/data/project/posts")
     );
@@ -167,18 +167,18 @@ describe("Builder - HTML", () => {
         "<div> <div><div>Fourth post</div><p>Test</p><p>Fourth blog test</p><p>04-01-2019</p><div>Third post</div><p>Test</p><p>Third blog test</p><p>03-01-2019</p></div></div>"
     };
 
-    const output = await htmlBuilder(store, fakeFile);
+    const output = await htmlBuilder(cache, fakeFile);
     expect(output).toMatchObject(result);
   });
 
   test("Recent posts - use partial metadata", async () => {
-    const store = new Store();
-    store.set("config", { recentPosts: 2 });
-    store.set("templates", {
+    const cache = new MemoryCache();
+    cache.set("config", { recentPosts: 2 });
+    cache.set("templates", {
       main: "<div>{{content}}</div>",
       recentPost: "<div>{{title}}</div><p>{{author}}</p>"
     });
-    store.set(
+    cache.set(
       "posts",
       await DirScanner.scanAndGetFiles("./test/data/project/posts")
     );
@@ -197,16 +197,16 @@ describe("Builder - HTML", () => {
         "<div> <div><div>Fourth post</div><p>Test</p><div>Third post</div><p>Test</p></div></div>"
     };
 
-    const output = await htmlBuilder(store, fakeFile);
+    const output = await htmlBuilder(cache, fakeFile);
     expect(output).toMatchObject(result);
   });
 
   test("No archive post template file", async () => {
-    const store = new Store();
-    store.set("templates", {
+    const cache = new MemoryCache();
+    cache.set("templates", {
       main: "<div>{{content}}</div>"
     });
-    store.set("posts", DirScanner.scanAndGetFiles("./test/data/project/posts"));
+    cache.set("posts", DirScanner.scanAndGetFiles("./test/data/project/posts"));
 
     const fakeFile: IFile = {
       name: "article.html",
@@ -216,18 +216,18 @@ describe("Builder - HTML", () => {
       raw: "<!--\n @template main \n--> <div>{{blog:archive}}</div>"
     };
 
-    return expect(htmlBuilder(store, fakeFile)).rejects.toThrow(
+    return expect(htmlBuilder(cache, fakeFile)).rejects.toThrow(
       new Error("There is no template for archive.")
     );
   });
 
   test("Archive", async () => {
-    const store = new Store();
-    store.set("templates", {
+    const cache = new MemoryCache();
+    cache.set("templates", {
       main: "<div>{{content}}</div>",
       archivePost: "<div>{{title}}</div><p>{{date}}</p>"
     });
-    store.set(
+    cache.set(
       "posts",
       await DirScanner.scanAndGetFiles("./test/data/project/posts")
     );
@@ -246,7 +246,7 @@ describe("Builder - HTML", () => {
         "<div> <div><div>Fourth post</div><p>04-01-2019</p><div>Third post</div><p>03-01-2019</p><div>Second post</div><p>02-01-2019</p><div>First post</div><p>01-01-2019</p></div></div>"
     };
 
-    const output = await htmlBuilder(store, fakeFile);
+    const output = await htmlBuilder(cache, fakeFile);
     expect(output).toMatchObject(result);
   });
 });