+++ test/unit/builder.test.ts
@@ -1,10 +1,10 @@
 import { Builder, IBuilders } from "../../src/domain/builder";
-import { Store } from "../../src/lib/store";
+import { MemoryCache } from "../../src/lib/cache";
 import { IFile } from "../../src/interfaces";
 
 describe("Builder", () => {
   test("Build file calls registered builder", async () => {
-    const store = new Store();
+    const cache = new MemoryCache();
     const fakeFile: IFile = {
       name: "test.html",
       extension: "html",
@@ -12,22 +12,22 @@ describe("Builder", () => {
       path: "",
       raw: "fake-file"
     };
-    const fakeBuilder = jest.fn(async (store, file) => {
+    const fakeBuilder = jest.fn(async (cache, file) => {
       return file;
     });
     const builders: IBuilders = {
       html: fakeBuilder
     };
 
-    const builder = new Builder(store, builders);
+    const builder = new Builder(cache, builders);
     const output = await builder.buildFile(fakeFile);
 
     expect(output).toMatchObject(fakeFile);
-    expect(fakeBuilder).toHaveBeenCalledWith(store, fakeFile);
+    expect(fakeBuilder).toHaveBeenCalledWith(cache, fakeFile);
   });
 
   test("Build file calls unregistered builder", async () => {
-    const store = new Store();
+    const cache = new MemoryCache();
     const fakeFile: IFile = {
       name: "test.css",
       extension: "css",
@@ -35,14 +35,14 @@ describe("Builder", () => {
       path: "",
       raw: "fake-file"
     };
-    const fakeBuilder = jest.fn(async (store, file) => {
+    const fakeBuilder = jest.fn(async (cache, file) => {
       return file;
     });
     const builders: IBuilders = {
       html: fakeBuilder
     };
 
-    const builder = new Builder(store, builders);
+    const builder = new Builder(cache, builders);
     const output = await builder.buildFile(fakeFile);
 
     expect(output).toMatchObject(fakeFile);