+++ CHANGELOG.md
@@ -1,13 +1,25 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 20-10-2019
+## [0.2.0] - 21-10-2019
+
 ### Added
+
+- Markdown support.
+- Build command supports output folder option and args validation.
+- Travis CI integration.
+- Test coverage.
+
+## [0.1.0] - 20-10-2019
+
+### Added
+
 - Uses generic file types only.
-   Template system.
+- Template system.
 - Snippet system.
 - Progressive builds.
 - CSS minification.